+++ schematics/src/upgrade/v2.ts
@@ -7,18 +7,24 @@ export function run(path) {
   const dir = p.resolve(process.cwd());
   path = p.join(dir, path, '/**/*');
   const htmlFiles = glob.sync(`${path}.html`);
+  const templateRegex = /<ng-template[^>]*transloco[^>]*>[^]+?<\/ng-template>/g;
   const structuralRegex = /<([a-zA-Z-]*)[^*>]*\*transloco=('|")\s*let\s+(?<varName>\w*)[^>]*\2>[^]+?<\/\1\s*>/g;
-  const templateKey = varName => new RegExp(`${varName}(?:(?:\\[(?:'|"))|\\.)([^}|:]*)`, 'g');
+  const templateKey = varName =>
+    new RegExp(`(?<rawKey>${varName}(?:(?:\\[(?:'|"))|\\.)(?:[^}|:]*))(?<param>[\\s\\w|:{}'"$&!?%@#^)(]*)}}`, 'g');
   for (const file of htmlFiles) {
     let str = fs.readFileSync(file).toString('utf8');
-    if (!str.includes('*transloco')) continue;
-    let result = structuralRegex.exec(str);
-    while (result) {
-      const [matchedStr] = result;
-      let { varName } = result.groups;
-      let scopeKeys = matchedStr.match(templateKey(varName));
-      scopeKeys &&
-        scopeKeys.forEach(rawKey => {
+    if (!str.includes('transloco')) continue;
+    [structuralRegex, templateRegex].forEach((rgx, index) => {
+      let containerSearch = rgx.exec(str);
+      while (containerSearch) {
+        const [matchedStr] = containerSearch;
+        let { varName } = index === 0 ? containerSearch.groups : matchedStr.match(/let-(?<varName>\w*)/).groups;
+        const keyRegex = templateKey(varName);
+        let newStructuralStr = matchedStr;
+        let keySearch = keyRegex.exec(matchedStr);
+        let hasMatches = !!keySearch;
+        while (keySearch) {
+          const { rawKey, param } = keySearch.groups;
           /** The raw key may contain square braces we need to align it to '.' */
           let [key, ...inner] = rawKey
             .trim()
@@ -26,12 +32,32 @@ export function run(path) {
             .replace(/'|"|\]|\?/g, '')
             .replace(`${varName}.`, '')
             .split('.');
-          if (inner.length) {
-            str = str.replace(rawKey, `${varName}['${key}.${inner.join('.')}']`);
+          let callEnd = ')';
+          const pipes = (param && param.split('|')) || [];
+          const [paramsPipe] = pipes.filter(pipe => pipe.includes('translocoParams'));
+          if (paramsPipe) {
+            const paramValue = paramsPipe.substring(paramsPipe.indexOf('{'), paramsPipe.lastIndexOf('}') + 1);
+            if (paramValue) {
+              callEnd = `, ${paramValue})`;
+            }
           }
-        });
-      result = structuralRegex.exec(str);
-    }
+          if (inner.length) {
+            key = `${key}.${inner.join('.')}`;
+          }
+          newStructuralStr = paramsPipe
+            ? newStructuralStr.replace(
+                `${rawKey}${param}`,
+                `${varName}('${key}'${callEnd} ${pipes.filter(pipe => !pipe.includes('translocoParams')).join('|')}`
+              )
+            : newStructuralStr.replace(rawKey, `${varName}('${key}'${callEnd}`);
+          keySearch = keyRegex.exec(matchedStr);
+        }
+        if (hasMatches) {
+          str = str.replace(matchedStr, newStructuralStr);
+        }
+        containerSearch = rgx.exec(str);
+      }
+    });
     fs.writeFileSync(file, str, { encoding: 'utf8' });
   }
   const modules = glob.sync(`${path}.module.ts`);